Avoid loading virtualbox driver on Jenkins slaves
Trying to destroy a VM before the machine is available loads the
virtualbox driver, VirtualBox isn't installed on our Jenkins slaves

+++ bosh-dev/lib/bosh/dev/stemcell_vm.rb
@@ -6,12 +6,13 @@ module Bosh::Dev
       @operating_system_name = options.fetch(:operating_system_name)
       @env = env
     end
-#
+
     def publish
       Rake::FileUtilsExt.sh <<-BASH
         set -eu
 
-        vagrant destroy #{vm_name} --force
+        cd bosh-stemcell
+        [ -e .vagrant/machines/remote/aws/id ] && vagrant destroy #{vm_name} --force
         vagrant up #{vm_name} --provider #{provider}
 
         time vagrant ssh -c "
